Index Look
I have managed to hide the links when you press buttons and added a H3 under the title, with bigger text H1 and lowered it a bit down to make it more centered. I also think that if you make the buttons text cleaner instead of cursive. Of course I could change the subtittle text. I am going to start playing with it a bit more. Let me know what you think.